fixed checkstyle issues, added missing mock and removed duplicate test
